requirements.txt
mosquitto==1.2.2
PyYAML==3.10
pyserial==2.7